Add Header.js if문 등 사용을 위해 header component 분리
